kv-r2-caches
