+++ fixtures/vitest-pool-workers-examples/kv-r2-caches/src/env.d.ts
@@ -1,4 +1,12 @@
-interface Env {
-	KV_NAMESPACE: KVNamespace;
-	R2_BUCKET: R2Bucket;
+/* eslint-disable */
+// Generated by Wrangler by running `wrangler types --no-include-runtime ./src/env.d.ts` (hash: e727226ca8c7f070ed97a0ad5637d454)
+declare namespace Cloudflare {
+	interface GlobalProps {
+		mainModule: typeof import("./index");
+	}
+	interface Env {
+		KV_NAMESPACE: KVNamespace;
+		R2_BUCKET: R2Bucket;
+	}
 }
+interface Env extends Cloudflare.Env {}
