--- fixtures/vitest-pool-workers-examples/kv-r2-caches/test/env.d.ts
@@ -1,6 +1,0 @@
-declare namespace Cloudflare {
-	interface Env {
-		KV_NAMESPACE: KVNamespace;
-		R2_BUCKET: R2Bucket;
-	}
-}
