+++ fixtures/vitest-pool-workers-examples/kv-r2-caches/test/kv.test.ts
@@ -1,14 +1,14 @@
-import { SELF } from "cloudflare:test";
+import { exports } from "cloudflare:workers";
 import { it } from "vitest";
 
 it("stores in KV namespace", async ({ expect }) => {
-	let response = await SELF.fetch("https://example.com/kv/key", {
+	let response = await exports.default.fetch("https://example.com/kv/key", {
 		method: "PUT",
 		body: "value",
 	});
 	expect(response.status).toBe(204);
 
-	response = await SELF.fetch("https://example.com/kv/key");
+	response = await exports.default.fetch("https://example.com/kv/key");
 	expect(response.status).toBe(200);
 	expect(await response.text()).toBe("value");
 });
